moves code to src
