+++ runConvertImperialToJsonl.js
@@ -1,6 +1,6 @@
 const fs = require('file-system');
 const path = require('path')
-const ConvertImperialToJsonl = require('./ConvertImperialToJsonl');
+const ConvertImperialToJsonl = require('./src/converters/ConvertImperialToJsonl');
 const imperialToJsonlConverter = new ConvertImperialToJsonl(fs, path);
 const convertedData = imperialToJsonlConverter.convertRawImperialDataToJsonl();
 module.exports = convertedData;
